requirements-sam2.txt
antlr4-python3-runtime==4.9.3
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decord==0.6.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
huggingface-hub==0.32.2
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
portalocker==3.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0

PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
urllib3==2.4.0